+++ src/resources/Namespaces/YamlUpload/SeparatorLine.js
@@ -1,5 +1,5 @@
 import React from 'react';
 
 export const SeparatorLine = ({ style }) => {
-  return <hr className={'yaml_resource_list__separation-line'} style={style} />;
+  return <hr className="yaml_resource_list__separation-line" style={style} />;
 };
