+++ src/resources/Namespaces/YamlUpload/SeparatorLine.scss
@@ -1,8 +1,3 @@
-.validate-resources {
-  display: flex;
-  align-items: center;
-}
-
 .yaml_resource_list {
   &__separation-line {
     width: 99%;
