+++ src/resources/Namespaces/YamlUpload/ValidationSwitch.scss
@@ -1,0 +1,4 @@
+.validate-resources {
+  display: flex;
+  align-items: center;
+}
